+++ CHANGELOG.md
@@ -23,6 +23,20 @@ If you like this project and find it useful, please consider giving it a star on
   <img src="bmc-button.svg" alt="Buy me a coffee" width="120">
 </a>
 
+## [1.3.2] - 2025-07-04
+
+### Added
+
+- [platform]: Add by default Rvc in mode server for Apple Home issue.
+
+### Changed
+
+- [package]: Updated dependencies.
+
+<a href="https://www.buymeacoffee.com/luligugithub">
+  <img src="bmc-button.svg" alt="Buy me a coffee" width="80">
+</a>
+
 ## [1.3.1] - 2025-07-04
 
 ### Added